--- a/ethcore/src/blooms/bloom_group.rs
+++ b/ethcore/src/blooms/bloom_group.rs
@@ -63,7 +63,7 @@ impl Decodable for BloomGroup {
 
 impl Encodable for BloomGroup {
 	fn rlp_append(&self, s: &mut RlpStream) {
-		Encodable::rlp_append(&self.blooms, s)
+		s.append_list(&self.blooms);
 	}
 }