--- a/ethcore/src/migrations/state/v7.rs
+++ b/ethcore/src/migrations/state/v7.rs
@@ -26,7 +26,7 @@ use util::migration::{Batch, Config, Error, Migration, SimpleMigration, Progress
 use util::sha3::Hashable;
 use std::sync::Arc;
 
-use rlp::{decode, Rlp, RlpStream, Stream, View};
+use rlp::{decode, Rlp, RlpStream, View};
 
 
 // attempt to migrate a key, value pair. None if migration not possible.
@@ -199,7 +199,7 @@ impl OverlayRecentV7 {
 							stream.begin_list(2).append(&k).append(&v);
 						}
 
-						stream.append(&deleted_keys);
+						stream.append_list(&deleted_keys);
 
 						// and insert it into the new database.
 						batch.insert(entry_key, stream.out(), dest)?;