--- a/ethcore/src/migrations/v9.rs
+++ b/ethcore/src/migrations/v9.rs
@@ -17,7 +17,7 @@
 
 //! This migration consolidates all databases into single one using Column Families.
 
-use rlp::{Rlp, RlpStream, View, Stream};
+use rlp::{Rlp, RlpStream, View};
 use util::kvdb::Database;
 use util::migration::{Batch, Config, Error, Migration, Progress};
 use std::sync::Arc;