--- a/util/rlp/src/impls.rs
+++ b/util/rlp/src/impls.rs
@@ -1,4 +1,4 @@
-use byteorder::{WriteBytesExt, BigEndian};
+use byteorder::{ByteOrder, BigEndian};
 use bigint::prelude::{Uint, U128, U256, H64, H128, H160, H256, H512, H520, H2048};
 use traits::Encodable;
 use stream::RlpStream;
@@ -55,7 +55,7 @@ macro_rules! impl_encodable_for_u {
 			fn rlp_append(&self, s: &mut RlpStream) {
 				let leading_empty_bytes = self.leading_zeros() as usize / 8;
 				let mut buffer = [0u8; $size];
-				(&mut buffer as &mut [u8]).$func::<BigEndian>(*self).expect("buffer.len() == sizeof(*self); qed");
+				BigEndian::$func(&mut buffer, *self);
 				s.encoder().encode_value(&buffer[leading_empty_bytes..]);
 			}
 		}