--- a/util/rlp/src/lib.rs
+++ b/util/rlp/src/lib.rs
@@ -71,7 +71,7 @@ use std::borrow::Borrow;
 use elastic_array::ElasticArray1024;
 
 pub use error::DecoderError;
-pub use traits::{Decoder, Decodable, View, Stream, Encodable, RlpEncodable, RlpDecodable, Compressible};
+pub use traits::{Decoder, Decodable, View, RlpEncodable, RlpDecodable, Compressible};
 pub use untrusted_rlp::{UntrustedRlp, UntrustedRlpIterator, PayloadInfo, Prototype};
 pub use rlpin::{Rlp, RlpIterator};
 pub use stream::RlpStream;