for now, it's only serialization refactor. it's still WIP, but wanted to know, if you like interface changes I made:

changes:

- removed `BytesEncodable` trait
- removed `Encodable` trait
- serialization do not use `ToBytes`
- `Encodable` is no longer implemented to `T where T: ToBytes`
- `RlpStream` does not implement `Stream` trait. All functions where moved to public interface. I find additional `Stream` import to be really annoying
- used `byteorder` for serialization
- new api method `append_list` which should be used to append list of encodable items
- new api method `append_internal` which should be used for wrapper types to encoding item without incrementing list elements counter
- serialization of lists preallocates 1 additional byte for payload. This allows us avoid some unnecessary shifts for short lists


benchmarks:

- **old**

```
test bench_stream_1000_empty_lists   ... bench:       9,839 ns/iter (+/- 1,207)
test bench_stream_nested_empty_lists ... bench:         223 ns/iter (+/- 23)
test bench_stream_u256_value         ... bench:         453 ns/iter (+/- 52)
test bench_stream_u64_value          ... bench:         119 ns/iter (+/- 16)
```

- **new (update)**

```
test bench_stream_1000_empty_lists   ... bench:       9,143 ns/iter (+/- 1,209)
test bench_stream_nested_empty_lists ... bench:         181 ns/iter (+/- 22)
test bench_stream_u256_value         ... bench:         344 ns/iter (+/- 53)
test bench_stream_u64_value          ... bench:          90 ns/iter (+/- 10)
```

so encoding integers is **almost 25% faster**