@gavofyork 2017-03-14
reduction of code size and number of traits is good.

addition of `_list` everywhere less so.