@gavofyork 2017-03-15
is it the case that the only underlying reason for the `append`/`append_list` API divergence is the need to distinguish between `Vec<u8>` as a single piece of data and as a list of numbers?