@gavofyork 2017-03-15
i seem to remember there was something we could do with traits, no? Have a `Listable` trait which is good for all types except `u8`?