@debris 2017-03-17
> i seem to remember there was something we could do with traits, no? Have a Listable trait which is good for all types except u8?

Nothing less complicated than what we had before (with at least 3 different traits) comes to my mind. And I think the solution from this pr is less complicated and error prone