@gavofyork 2017-03-19
well the fact i now have two different functions to call in order to do the same conceptual thing, depending on some arbitrary distinction between datatypes, adds to the complexity.

that said, i won't block it if y'all prefer this way of doing it.